# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for Feedback service",
        "title": "Tourmate - Feedback Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/feedbacks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a paginated list of feedbacks with optional filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feedbacks"
                ],
                "summary": "Get all feedbacks",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page_number",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search keyword",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter property",
                        "name": "filter_prop",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (ASC or DESC)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Rating",
                        "name": "rating",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "The owner ID of this feedback",
                        "name": "customer_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Tour guide ID",
                        "name": "tour_guide_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Invoice ID",
                        "name": "invoice_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.PaginationDataResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid data. Please try again.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feedbacks"
                ],
                "summary": "Update an existing feedback",
                "parameters": [
                    {
                        "description": "Update Feedback Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateFeedbackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid data. Please try again.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    },
                    "401": {
                        "description": "You have no rights to access this action.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    },
                    "500": {
                        "description": "There is something wrong in the system during the process. Please try again later.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feedbacks"
                ],
                "summary": "Create a new feedback",
                "parameters": [
                    {
                        "description": "Create Feedback Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateFeedbackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid data. Please try again.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    },
                    "401": {
                        "description": "You have no rights to access this action.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    },
                    "500": {
                        "description": "There is something wrong in the system during the process. Please try again later.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feedbacks"
                ],
                "summary": "Remove a feedback",
                "parameters": [
                    {
                        "description": "Remove Feedback Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RemoveFeedbackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid data. Please try again.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    },
                    "401": {
                        "description": "You have no rights to access this action.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    },
                    "500": {
                        "description": "There is something wrong in the system during the process. Please try again later.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/feedbacks/user/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve feedbacks filtered by customer ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feedbacks"
                ],
                "summary": "Get feedbacks by user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Customer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page_number",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search keyword",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter property",
                        "name": "filter_prop",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (ASC or DESC)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Rating",
                        "name": "rating",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Tour guide ID",
                        "name": "tour_guide_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Invoice ID",
                        "name": "invoice_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.PaginationDataResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid data. Please try again.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/feedbacks/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feedbacks"
                ],
                "summary": "Get a single feedback by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Feedback ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Feedback"
                        }
                    },
                    "400": {
                        "description": "Invalid data. Please try again.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entity.Feedback": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_date": {
                    "type": "string"
                },
                "customer_id": {
                    "type": "integer"
                },
                "feedback_id": {
                    "type": "integer"
                },
                "invoice_id": {
                    "type": "integer"
                },
                "is_deleted": {
                    "type": "boolean"
                },
                "rating": {
                    "type": "integer"
                },
                "tour_guide_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "request.CreateFeedbackRequest": {
            "type": "object",
            "required": [
                "customer_id",
                "invoice_id",
                "rating",
                "tour_guide_id"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "customer_id": {
                    "type": "integer"
                },
                "invoice_id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "integer"
                },
                "tour_guide_id": {
                    "type": "integer"
                }
            }
        },
...
        "request.RemoveFeedbackRequest": {
            "type": "object",
            "required": [
                "actor_id",
                "feedback_id"
            ],
            "properties": {
                "actor_id": {
                    "type": "integer"
                },
                "feedback_id": {
                    "type": "integer"
                }
            }
        },
        "request.UpdateFeedbackRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                },
                "request": {
                    "$ref": "#/definitions/request.RemoveFeedbackRequest"
                }
            }
        },
...
        "response.MessageApiResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "response.PaginationDataResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "has_next": {
                    "type": "boolean"
                },
                "has_previous": {
                    "type": "boolean"
                },
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}